% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labels.R
\name{asLabels}
\alias{asLabels}
\title{Utilizar etiquetas de datos de la variable}
\usage{
asLabels(var)
}
\arguments{
\item{var}{variable o vector de datos que contiene las etiquetas a ser empleadas}
}
\value{
vector de datos con los valores cambiados por etiquetas
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}
}
\details{
Reemplaza los valores de la variable por las etiquetas de esta
}
\examples{
\dontrun{
  tbl$var <- asLabels(tbl$var)
}
}
