% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_dict.R
\name{browse_dict}
\alias{browse_dict}
\title{Navegar diccionario}
\usage{
browse_dict(dict)
}
\arguments{
\item{dict}{diccionario de base de datos}
}
\value{
una interfaz web con los datos contenidos en el diccionario suministrado
}
\description{
Permite navegar el diccionario de las encuestas en una interfaz web donde
se pueden consultar el nombre y etiquitada de las variables, así como las
etiquetas de los datos.
}
\examples{
\dontrun{
  browse_dict(enhogar::dict)
}
}
