% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endomer-package.R
\docType{package}
\name{endomer-package}
\alias{endomer}
\alias{endomer-package}
\title{Metapaquete para las encuestas ENFT, ENCFT, ENHOGAR y ENGIH
Declara versiones minimas compatibles, carga las cuatro encuestas con library()
y ofrece informes de versiones sin instalar ni actualizar automaticamente.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Reune las interfaces para ENFT, ENCFT, ENHOGAR y ENGIH mediante versiones minimas declaradas y carga conjunta. Permite inspeccionar versiones instaladas, contrastarlas con el catalogo incluido o un indice de repositorio explicito y preparar informes de actualizacion sin modificar las bibliotecas de R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/endomer-py/endomer}
  \item \url{https://endomer-py.github.io/endomer/}
  \item Report bugs at \url{https://github.com/endomer-py/endomer/issues}
}

}
\author{
\strong{Maintainer}: Daniel E. de la Rosa \email{me@dnldelarosa.com}

Other contributors:
\itemize{
  \item Adatar [copyright holder]
}

}
\keyword{internal}
