% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{endomer_packages}
\alias{endomer_packages}
\title{Paquetes de encuestas incluidos en endomer}
\usage{
endomer_packages()
}
\value{
data.frame con package, required, repository y documentation.
}
\description{
El catalogo versionado declara las versiones minimas de las cuatro encuestas.
No consulta Internet ni representa las versiones mas recientes de GitHub.
}
\examples{
endomer_packages()
}
