% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labels.R
\name{setLabels}
\alias{setLabels}
\title{Etiquetas de datos}
\usage{
setLabels(tbl, vars = NULL, pattern)
}
\arguments{
\item{tbl}{Conexión a base de datos o dataframe con los datos de la ENCFT.}

\item{vars}{Si especificado, solo se asignaran las etiquetas a esas variables.}

\item{pattern}{patrón inicial del nombre de las funciones del paquete}
}
\value{
Conexión a base de datos o dataframe según input.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\details{
Asigna etiquetas de datos a todas las variables que aplica.
}
\examples{
\dontrun{
  encft <- setLabels(tbl = encft, pattern = "ftc")
}
}
