% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labels.R
\name{useLabels}
\alias{useLabels}
\title{Utilizar etiquetas de datos de todas las variables}
\usage{
useLabels(tbl, vars = NULL)
}
\arguments{
\item{tbl}{Conexión a base de datos o dataframe con los datos}

\item{vars}{Si es especificado, solo estas variables serán asignadas como
etiquetas de datos}
}
\value{
Conexión a base de datos o dataframe según input
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\examples{
\dontrun{
  encft <- useLabels(encft)
}
}
