% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{endomer_deps}
\alias{endomer_deps}
\title{Lista todas las dependencias de ENDOMER}
\usage{
endomer_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{Si \code{TRUE}, incluso se listan las dependencias de las
dependencias de ENDOMER.}

\item{repos}{Los repositorios que se utilizarán para verificar las actualizaciones.
Por defecto se toma \code{getOption("repos")}.}
}
\description{
Lista todas las dependencias de ENDOMER
}
