% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{endomer_update}
\alias{endomer_update}
\title{Actualiza los paquetes ENDOMER}
\usage{
endomer_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{Si \code{TRUE}, incluso se listan las dependencias de las
dependencias de ENDOMER.}

\item{repos}{Los repositorios que se utilizarán para verificar las actualizaciones.
Por defecto se toma \code{getOption("repos")}.}
}
\description{
Se valida si hay actualizaciones disponibles para los paquetes ENDOMER, y se
pregunta al usuario si desea actualizarlos.
}
\examples{
\dontrun{
endomer_update()
}
}
