% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labels.R
\name{setLabels}
\alias{setLabels}
\title{Asigna etiquetas de datos a las variables especificadas}
\usage{
setLabels(tbl, vars = NULL, dict, ignore_case = FALSE)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}

\item{vars}{\link{character}: Si especificado, solo se asignaran las etiquetas a esas variables.}

\item{dict}{\link{data.frame}: Diccionario con todas las etiquetas de datos a utilizar}

\item{ignore_case}{\link{logical}: Indicate if case sensitive should be ignored.}
}
\value{
Los datos introducidos en el argumento \code{tbl} pero con etiquetas de datos
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Esta función permite asignar etiquetas de datos a las variables de un data.frame
o equivalente, a partir de un diccionario suministrado. En tal sentido,
puedes utilizar esta función suministrando un diccionario peronal o utilizar
las funciones equivalentes en los paquetes ENDOMER que traen un diccionario
predefinido.
}
\examples{
\dontrun{
  enft <- data.frame(S2_P8 = c(1, 2))
  str(enft)
  str(setLabels(enft, dict = dict))
}
}
\seealso{
Etiquetas de datos \code{vignette("labels", package = "endomer")}
}
